+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.3.0] - 2020-09-11
+
+### Added
+
+- Input for specifying the type of the project’s version file.
+- Input for specifying path to the project’s version file.
+- Support for text version files.
+
+### Changed
+
+- Place of the version upload to POST action.
+
 ## [0.2.3] - 2020-09-11
 
 ### Added
@@ -46,7 +58,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Build scripts and package information using Node.js.
 - Functionality for handling the version number increments in projects that use Node.js.
 
-[unreleased]: https://github.com/anttikivi/maintain-revision/compare/v0.2.3...HEAD
+[unreleased]: https://github.com/anttikivi/maintain-revision/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/anttikivi/maintain-revision/compare/v0.2.3...v0.3.0
 [0.2.3]: https://github.com/anttikivi/maintain-revision/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/anttikivi/maintain-revision/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/anttikivi/maintain-revision/compare/v0.2.0...v0.2.1